Changed to lookup table and using integers

+++ src/fft.h
@@ -1,4 +1,4 @@
 #ifdef fft_lib
 #define fft_lib
-void fft(double* data,double* result,int N);
+void fft(int* data,int* result,int N);
 #endif
